refactor(dead): split the dead-code package into cohesive files
internal/dead/dead.go (1063 lines) held FindDead, buildFacts, the read*
meta-readers, the predicates, and the candidate SQL in one wall. Split it
into orchestration-only dead.go (FindDead + buildFacts + defaultArbiter)
plus three cohesive files, each answering one question:

- meta_readers.go: the harvested-name readers (sense_meta loaders)
- predicates.go:   the is*/exclude* symbol classifiers and filter types
- queries.go:      the index SQL helpers

Collapse the twelve flat per-feature read*Names wrappers (each a pure
readStringSetMeta(key) call whose doc duplicated its Facts field) into
direct readStringSetMeta(ctx, db, "<key>") calls at their buildFacts
assignments, naming each sense_meta key once beside its field.

Decompose the two over-threshold functions: extract loadChildrenByParent
from findLiveContainers, and split isInTestFile into inTestDir /
inFixtureDir / hasTestFileSuffix. The lone //nolint:gocyclo is retired.

Behavior is unchanged: the maket dead-code query is byte-identical before
and after (dead 2, possibly_dead 3815, total 10809), verified on the
existing index with no rescan.
